Refactor product model and CSV handling: change price type to double, introduce ProductStatus enum for product status management, update CSV file path, and rename user response package.
