+++ backend/src/main/java/com/nlu/WebThuongMai/dto/request/productReq/ProductRequest.java
@@ -16,7 +16,7 @@ import java.util.Set;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ProductRequest {
     String name;
-    String price;
+    Double price;
     String status;
     Set<Long> categoryIds;
     String description;
