+++ backend/src/main/java/com/nlu/WebThuongMai/enums/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.nlu.WebThuongMai.enums;
+
+public enum ProductStatus {
+    ACTIVE, INACTIVE, DELETED
+}
