+++ backend/src/main/resources/db/migration/data/categories.sql
@@ -1,1 +1,8 @@
-;
+insert into public.categories (category_id, description, image_url, name)
+values  (1, null, null, 'Bi-a'),
+        (2, null, null, 'Bóng chuyền'),
+        (3, null, null, 'Pickleball'),
+        (4, null, null, 'Chạy bộ'),
+        (5, null, null, 'Bóng đá & Futsal'),
+        (6, null, null, 'Bóng rổ'),
+        (7, null, null, 'Cầu lông');
