+++ backend/src/main/resources/db/migration/data/product_categories.sql
@@ -1,1 +1,257 @@
-;
+insert into public.product_categories (category_id, product_id)
+values  (7, 1),
+        (7, 2),
+        (7, 3),
+        (7, 4),
+        (7, 5),
+        (7, 6),
+        (7, 7),
+        (7, 8),
+        (7, 9),
+        (7, 10),
+        (7, 11),
+        (7, 12),
+        (7, 13),
+        (7, 14),
+        (7, 15),
+        (7, 16),
+        (7, 17),
+        (7, 18),
+        (7, 19),
+        (7, 20),
+        (7, 21),
+        (7, 22),
+        (7, 23),
+        (7, 24),
+        (7, 25),
+        (7, 26),
+        (7, 27),
+        (7, 28),
+        (7, 29),
+        (7, 30),
+        (7, 31),
+        (7, 32),
+        (6, 33),
+        (6, 34),
+        (6, 35),
+        (6, 36),
+        (6, 37),
+        (6, 38),
+        (6, 39),
+        (6, 40),
+        (6, 41),
+        (6, 42),
+        (6, 43),
+        (6, 44),
+        (6, 45),
+        (6, 46),
+        (6, 47),
+        (6, 48),
+        (6, 49),
+        (6, 50),
+        (6, 51),
+        (6, 52),
+        (6, 53),
+        (6, 54),
+        (6, 55),
+        (6, 56),
+        (6, 57),
+        (6, 58),
+        (6, 59),
+        (6, 60),
+        (6, 61),
+        (6, 62),
+        (6, 63),
+        (6, 64),
+        (1, 65),
+        (1, 66),
+        (1, 67),
+        (1, 68),
+        (1, 69),
+        (1, 70),
+        (1, 71),
+        (1, 72),
+        (1, 73),
+        (1, 74),
+        (1, 75),
+        (1, 76),
+        (1, 77),
+        (1, 78),
+        (1, 79),
+        (1, 80),
+        (1, 81),
+        (1, 82),
+        (1, 83),
+        (1, 84),
+        (1, 85),
+        (1, 86),
+        (1, 87),
+        (1, 88),
+        (1, 89),
+        (1, 90),
+        (1, 91),
+        (1, 92),
+        (1, 93),
+        (1, 94),
+        (1, 95),
+        (1, 96),
+        (3, 97),
+        (3, 98),
+        (3, 99),
+        (3, 100),
+        (3, 101),
+        (3, 102),
+        (3, 103),
+        (3, 104),
+        (3, 105),
+        (3, 106),
+        (3, 107),
+        (3, 108),
+        (3, 109),
+        (3, 110),
+        (3, 111),
+        (3, 112),
+        (3, 113),
+        (3, 114),
+        (3, 115),
+        (3, 116),
+        (3, 117),
+        (3, 118),
+        (3, 119),
+        (3, 120),
+        (3, 121),
+        (3, 122),
+        (3, 123),
+        (3, 124),
+        (3, 125),
+        (3, 126),
+        (3, 127),
+        (3, 128),
+        (5, 129),
+        (5, 130),
+        (5, 131),
+        (5, 132),
+        (5, 133),
+        (5, 134),
+        (5, 135),
+        (5, 136),
+        (5, 137),
+        (5, 138),
+        (5, 139),
+        (5, 140),
+        (5, 141),
+        (5, 142),
+        (5, 143),
+        (5, 144),
+        (5, 145),
+        (5, 146),
+        (5, 147),
+        (5, 148),
+        (5, 149),
+        (5, 150),
+        (5, 151),
+        (5, 152),
+        (5, 153),
+        (5, 154),
+        (5, 155),
+        (5, 156),
+        (5, 157),
+        (5, 158),
+        (5, 159),
+        (5, 160),
+        (2, 161),
+        (2, 162),
+        (2, 163),
+        (2, 164),
+        (2, 165),
+        (2, 166),
+        (2, 167),
+        (2, 168),
+        (2, 169),
+        (2, 170),
+        (2, 171),
+        (2, 172),
+        (2, 173),
+        (2, 174),
+        (2, 175),
+        (2, 176),
+        (2, 177),
+        (2, 178),
+        (2, 179),
+        (2, 180),
+        (2, 181),
+        (2, 182),
+        (2, 183),
+        (2, 184),
+        (2, 185),
+        (2, 186),
+        (2, 187),
+        (2, 188),
+        (2, 189),
+        (2, 190),
+        (2, 191),
+        (2, 192),
+        (4, 193),
+        (4, 194),
+        (4, 195),
+        (4, 196),
+        (4, 197),
+        (4, 198),
+        (4, 199),
+        (4, 200),
+        (4, 201),
+        (4, 202),
+        (4, 203),
+        (4, 204),
+        (4, 205),
+        (4, 206),
+        (4, 207),
+        (4, 208),
+        (4, 209),
+        (4, 210),
+        (4, 211),
+        (4, 212),
+        (4, 213),
+        (4, 214),
+        (4, 215),
+        (4, 216),
+        (4, 217),
+        (4, 218),
+        (4, 219),
+        (4, 220),
+        (4, 221),
+        (4, 222),
+        (4, 223),
+        (4, 224),
+        (4, 225),
+        (4, 226),
+        (4, 227),
+        (4, 228),
+        (4, 229),
+        (4, 230),
+        (4, 231),
+        (4, 232),
+        (4, 233),
+        (4, 234),
+        (4, 235),
+        (4, 236),
+        (4, 237),
+        (4, 238),
+        (4, 239),
+        (4, 240),
+        (4, 241),
+        (4, 242),
+        (4, 243),
+        (4, 244),
+        (4, 245),
+        (4, 246),
+        (4, 247),
+        (4, 248),
+        (4, 249),
+        (4, 250),
+        (4, 251),
+        (4, 252),
+        (4, 253),
+        (4, 254),
+        (4, 255),
+        (4, 256);
