+++ backend/src/main/resources/db/migration/data/purchase_orders.sql
@@ -1,2 +1,1 @@
-insert into public.purchase_orders (total_price, created_at, purchase_order_id, suppler_id, status)
-values  (1500000.00, '2025-05-26 11:03:22.325191', 1, 5, 'PENDING');
+;
