--- backend/src/main/resources/db/migration/schema/myDatabase.sql
@@ -1,255 +1,0 @@
-create table user_coupon
-(
-    coupon_id bigint not null,
-    user_id   bigint not null,
-    primary key (coupon_id, user_id)
-);
-
-alter table user_coupon
-    owner to myuser;
-
-create table images_product
-(
-    product_id       bigint,
-    product_image_id bigint generated by default as identity
-        primary key,
-    image            varchar(255),
-    name             varchar(255)
-);
-
-alter table images_product
-    owner to myuser;
-
-create table categories
-(
-    category_id bigint generated by default as identity
-        primary key,
-    description varchar(255),
-    image_url   varchar(255),
-    name        varchar(255)
-);
-
-alter table categories
-    owner to myuser;
-
-create table coupons
-(
-    discount_percentage double precision not null,
-    limit_user          integer          not null,
-    coupon_id           bigint generated by default as identity
-        primary key,
-    created_at          timestamp(6),
-    expiration_date     timestamp(6),
-    code                varchar(255)     not null
-        unique,
-    description         varchar(255)
-);
-
-alter table coupons
-    owner to myuser;
-
-create table invalidated_token
-(
-    expiry_time timestamp(6),
-    id          varchar(255) not null
-        primary key
-);
-
-alter table invalidated_token
-    owner to myuser;
-
-create table products
-(
-    product_id  bigint generated by default as identity
-        primary key,
-    description varchar(255),
-    name        varchar(255) not null,
-    price       varchar(255),
-    status      varchar(255)
-);
-
-alter table products
-    owner to myuser;
-
-create table inventories
-(
-    quantity     integer not null,
-    inventory_id bigint generated by default as identity
-        primary key,
-    product_id   bigint  not null
-        constraint fk8drmqyx629j3oo8ct9jnc5y3y
-            references products
-);
-
-alter table inventories
-    owner to myuser;
-
-create table product_categories
-(
-    category_id bigint not null
-        constraint fkd112rx0alycddsms029iifrih
-            references categories,
-    product_id  bigint not null
-        constraint fklda9rad6s180ha3dl1ncsp8n7
-            references products,
-    primary key (category_id, product_id)
-);
-
-alter table product_categories
-    owner to myuser;
-
-create table product_images
-(
-    product_id       bigint
-        constraint fkqnq71xsohugpqwf3c9gxmsuy
-            references products,
-    product_image_id bigint generated by default as identity
-        primary key,
-    image            varchar(255),
-    name             varchar(255)
-);
-
-alter table product_images
-    owner to myuser;
-
-create table suppliers
-(
-    supplier_id  bigint generated by default as identity
-        primary key,
-    address      varchar(255),
-    contact_name varchar(255) not null,
-    description  varchar(255),
-    email        varchar(255) not null,
-    phone        varchar(255) not null
-);
-
-alter table suppliers
-    owner to myuser;
-
-create table purchase_orders
-(
-    total_price       numeric(38, 2),
-    created_at        timestamp(6),
-    purchase_order_id bigint generated by default as identity
-        primary key,
-    suppler_id        bigint not null
-        constraint fk2vni0n2dwgf1uc19gl6v6lsur
-            references suppliers,
-    status            varchar(255)
-        constraint purchase_orders_status_check
-            check ((status)::text = ANY
-                   ((ARRAY ['PENDING'::character varying, 'RECEIVED'::character varying, 'CANCELLED'::character varying])::text[]))
-);
-
-alter table purchase_orders
-    owner to myuser;
-
-create table purchase_order_items
-(
-    quantity               integer,
-    unit_price             numeric(38, 2),
-    product_id             bigint not null
-        constraint fks16e5vrvsp8alu0xp8m3a2ol5
-            references products,
-    purchase_order_id      bigint not null
-        constraint fko3yj8ocbw2kav38548t22hgh8
-            references purchase_orders,
-    purchase_order_item_id bigint generated by default as identity
-        primary key
-);
-
-alter table purchase_order_items
-    owner to myuser;
-
-create table users
-(
-    birthday      date,
-    created_at    timestamp(6),
-    updated_at    timestamp(6),
-    user_id       bigint generated by default as identity
-        primary key,
-    avatar        varchar(2000),
-    address       varchar(255),
-    auth_provider varchar(255)
-        constraint users_auth_provider_check
-            check ((auth_provider)::text = ANY
-                   ((ARRAY ['LOCAL'::character varying, 'FACEBOOK'::character varying, 'GOOGLE'::character varying])::text[])),
-    coupons       varchar(255),
-    email         varchar(255)
-        unique,
-    full_name     varchar(255),
-    gender        varchar(255)
-        constraint users_gender_check
-            check ((gender)::text = ANY
-                   ((ARRAY ['MALE'::character varying, 'FEMALE'::character varying, 'OTHER'::character varying])::text[])),
-    password      varchar(255),
-    phone         varchar(255)
-        unique,
-    provider_id   varchar(255),
-    role          varchar(255)
-        constraint users_role_check
-            check ((role)::text = ANY ((ARRAY ['ADMIN'::character varying, 'USER'::character varying])::text[])),
-    status        varchar(255),
-    username      varchar(255) not null
-        unique
-);
-
-alter table users
-    owner to myuser;
-
-create table orders
-(
-    discounted_price numeric(38, 2),
-    total_price      numeric(38, 2),
-    created_date     timestamp(6),
-    order_id         bigint generated by default as identity
-        primary key,
-    user_id          bigint not null
-        constraint fk32ql8ubntj5uh44ph9659tiih
-            references users,
-    note             varchar(255),
-    status           varchar(255)
-        constraint orders_status_check
-            check ((status)::text = ANY
-                   ((ARRAY ['PENDING'::character varying, 'SHIPPED'::character varying, 'DELIVERED'::character varying, 'CANCELLED'::character varying])::text[]))
-);
-
-alter table orders
-    owner to myuser;
-
-create table order_items
-(
-    price         numeric(38, 2),
-    quantity      integer,
-    order_id      bigint not null
-        constraint fkbioxgbv59vetrxe0ejfubep1w
-            references orders,
-    order_item_id bigint generated by default as identity
-        primary key,
-    product_id    bigint not null
-        constraint fkocimc7dtr037rh4ls4l95nlfi
-            references products
-);
-
-alter table order_items
-    owner to myuser;
-
-create table reviews
-(
-    rating     integer,
-    created_at timestamp(6),
-    product_id bigint not null
-        constraint fkpl51cejpw4gy5swfar8br9ngi
-            references products,
-    review_id  bigint generated by default as identity
-        primary key,
-    user_id    bigint not null
-        constraint fkcgy7qjc1r99dp117y9en6lxye
-            references users,
-    content    varchar(255)
-);
-
-alter table reviews
-    owner to myuser;
-
-
