--- backend/src/main/resources/db/migration/update-schema-v2.sql
@@ -1,182 +1,0 @@
-CREATE TABLE categories
-(
-    category_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    name        VARCHAR(255),
-    description VARCHAR(255),
-    CONSTRAINT pk_categories PRIMARY KEY (category_id)
-);
-
-CREATE TABLE coupons
-(
-    coupon_id           BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    code                VARCHAR(255)                            NOT NULL,
-    discount_percentage DOUBLE PRECISION                        NOT NULL,
-    description         VARCHAR(255),
-    limit_user          INTEGER                                 NOT NULL,
-    created_at          TIMESTAMP WITHOUT TIME ZONE,
-    expiration_date     TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_coupons PRIMARY KEY (coupon_id)
-);
-
-CREATE TABLE invalidated_token
-(
-    id          VARCHAR(255) NOT NULL,
-    expiry_time TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_invalidated_token PRIMARY KEY (id)
-);
-
-CREATE TABLE inventories
-(
-    inventory_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    product_id   BIGINT                                  NOT NULL,
-    quantity     INTEGER                                 NOT NULL,
-    CONSTRAINT pk_inventories PRIMARY KEY (inventory_id)
-);
-
-CREATE TABLE order_items
-(
-    order_item_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    product_id    BIGINT                                  NOT NULL,
-    order_id      BIGINT                                  NOT NULL,
-    quantity      INTEGER,
-    price         DECIMAL,
-    CONSTRAINT pk_order_items PRIMARY KEY (order_item_id)
-);
-
-CREATE TABLE orders
-(
-    order_id         BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    user_id          BIGINT                                  NOT NULL,
-    total_price      DECIMAL,
-    discounted_price DECIMAL,
-    status           VARCHAR(255),
-    note             VARCHAR(255),
-    created_date     TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_orders PRIMARY KEY (order_id)
-);
-
-CREATE TABLE product_images
-(
-    product_image_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    name             VARCHAR(255),
-    image            VARCHAR(255),
-    product_id       BIGINT                                  NOT NULL,
-    CONSTRAINT pk_product_images PRIMARY KEY (product_image_id)
-);
-
-CREATE TABLE products
-(
-    product_id  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    name        VARCHAR(255)                            NOT NULL,
-    price       VARCHAR(255),
-    description VARCHAR(255),
-    status      VARCHAR(255),
-    category_id BIGINT                                  NOT NULL,
-    CONSTRAINT pk_products PRIMARY KEY (product_id)
-);
-
-CREATE TABLE purchase_order_items
-(
-    purchase_order_item_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    purchase_order_id      BIGINT                                  NOT NULL,
-    product_id             BIGINT                                  NOT NULL,
-    quantity               INTEGER,
-    unit_price             DECIMAL,
-    CONSTRAINT pk_purchase_order_items PRIMARY KEY (purchase_order_item_id)
-);
-
-CREATE TABLE purchase_orders
-(
-    purchase_order_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    suppler_id        BIGINT                                  NOT NULL,
-    created_at        TIMESTAMP WITHOUT TIME ZONE,
-    status            VARCHAR(255),
-    total_price       DECIMAL,
-    CONSTRAINT pk_purchase_orders PRIMARY KEY (purchase_order_id)
-);
-
-CREATE TABLE reviews
-(
-    review_id  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    content    VARCHAR(255),
-    rating     INTEGER,
-    created_at TIMESTAMP WITHOUT TIME ZONE,
-    user_id    BIGINT                                  NOT NULL,
-    product_id BIGINT                                  NOT NULL,
-    CONSTRAINT pk_reviews PRIMARY KEY (review_id)
-);
-
-CREATE TABLE suppliers
-(
-    supplier_id  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    contact_name VARCHAR(255)                            NOT NULL,
-    phone        VARCHAR(255)                            NOT NULL,
-    email        VARCHAR(255)                            NOT NULL,
-    address      VARCHAR(255),
-    description  VARCHAR(255),
-    CONSTRAINT pk_suppliers PRIMARY KEY (supplier_id)
-);
-
-CREATE TABLE users
-(
-    user_id       BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    username      VARCHAR(255)                            NOT NULL,
-    password      VARCHAR(255),
-    email         VARCHAR(255)                            NOT NULL,
-    phone         VARCHAR(255),
-    auth_provider SMALLINT,
-    provider_id   BIGINT,
-    full_name     VARCHAR(255),
-    avatar        VARCHAR(2000),
-    birthday      date,
-    address       VARCHAR(255),
-    gender        VARCHAR(255),
-    role          VARCHAR(255),
-    status        VARCHAR(255),
-    coupons       VARCHAR(255),
-    created_at    TIMESTAMP WITHOUT TIME ZONE,
-    updated_at    TIMESTAMP WITHOUT TIME ZONE,
-    CONSTRAINT pk_users PRIMARY KEY (user_id)
-);
-
-ALTER TABLE coupons
-    ADD CONSTRAINT uc_coupons_code UNIQUE (code);
-
-ALTER TABLE users
-    ADD CONSTRAINT uc_users_email UNIQUE (email);
-
-ALTER TABLE users
-    ADD CONSTRAINT uc_users_phone UNIQUE (phone);
-
-ALTER TABLE users
-    ADD CONSTRAINT uc_users_username UNIQUE (username);
-
-ALTER TABLE inventories
-    ADD CONSTRAINT FK_INVENTORIES_ON_PRODUCT FOREIGN KEY (product_id) REFERENCES products (product_id);
-
-ALTER TABLE orders
-    ADD CONSTRAINT FK_ORDERS_ON_USER FOREIGN KEY (user_id) REFERENCES users (user_id);
-
-ALTER TABLE order_items
-    ADD CONSTRAINT FK_ORDER_ITEMS_ON_ORDER FOREIGN KEY (order_id) REFERENCES orders (order_id);
-
-ALTER TABLE order_items
-    ADD CONSTRAINT FK_ORDER_ITEMS_ON_PRODUCT FOREIGN KEY (product_id) REFERENCES products (product_id);
-
-ALTER TABLE products
-    ADD CONSTRAINT FK_PRODUCTS_ON_CATEGORY FOREIGN KEY (category_id) REFERENCES categories (category_id);
-
-ALTER TABLE purchase_orders
-    ADD CONSTRAINT FK_PURCHASE_ORDERS_ON_SUPPLER FOREIGN KEY (suppler_id) REFERENCES suppliers (supplier_id);
-
-ALTER TABLE purchase_order_items
-    ADD CONSTRAINT FK_PURCHASE_ORDER_ITEMS_ON_PRODUCT FOREIGN KEY (product_id) REFERENCES products (product_id);
-
-ALTER TABLE purchase_order_items
-    ADD CONSTRAINT FK_PURCHASE_ORDER_ITEMS_ON_PURCHASE_ORDER FOREIGN KEY (purchase_order_id) REFERENCES purchase_orders (purchase_order_id);
-
-ALTER TABLE reviews
-    ADD CONSTRAINT FK_REVIEWS_ON_PRODUCT FOREIGN KEY (product_id) REFERENCES products (product_id);
-
-ALTER TABLE reviews
-    ADD CONSTRAINT FK_REVIEWS_ON_USER FOREIGN KEY (user_id) REFERENCES users (user_id);
